meson: improve log path trimming
This is still from wlroots' meson code. I saw this was changed when
taking a look through the 0.19.0 milestone and it seemed like a nice
improvement.

The `fs.relative_to` method was added in Meson 1.3.0 which released in
November 2023, but that's close to a year ago at this point and most
distros should probably have it packaged.