server: count buffer locks and references
I'm not sure that this is totally correct but I'll have to see how it
works out (especially on different compositors).

Going forward, there will be at least one other piece of code which
needs to interact with server_buffers (that being the OpenGL
mirror/projector code.) Reference counting is needed to keep the backing
buffer data alive and lock counting is needed to ensure that
wl_buffer.release is not sent early.