config, lua, wrap: add mirror function
This allows the user to create a "mirror", which copies an arbitrary
subsection of the Minecraft window and displays it somewhere else within
the waywall window.

The function returns an object which, once garbage collected, causes the
mirror object to be destroyed. This ensures that the C objects which
back the mirror object cannot leak.