config, scene: create new scene_object API
This new method of handling scene objects (mirrors, images, text) will
allow for objects to be reordered on the Z axis arbitrarily.

Currently, Z ordering is just the order in which scene objects are
created, which is a breaking change (previously it was images in the
order of creation, then mirrors, then text) but I will try to replicate
the old behavior by default at least somewhat.