config: ignore caps/numlock when matching keybinds
This has tripped up a few people. I think it's fairly sensible to make
this happen automatically.

The only possible breakage I can think of is if somebody has a keybind
for e.g. `X` but *not* also `Caps-X`, but wants caps/numlock to prevent
the X action from running. This change would allow the X action to run.

This can be solved by just adding another bind for `Caps-X`, which is
more specific and would thus get matched first.