config, scene: add depth to scene objects (#20)
* config, scene: create new scene_object API

This new method of handling scene objects (mirrors, images, text) will
allow for objects to be reordered on the Z axis arbitrarily.

Currently, Z ordering is just the order in which scene objects are
created, which is a breaking change (previously it was images in the
order of creation, then mirrors, then text) but I will try to replicate
the old behavior by default at least somewhat.

* config, scene: add set_depth function

This is an absolute disaster at the moment but it does work. Need to add
depth options to all of the object constructors next.

* scene: fix object_sort

An object with depth n+1 needs to be inserted past all objects of depth
n, not just past the first object.

* scene: draw negative depth objects "behind" game

The stencil buffer is used to mask out the instance so that nothing can
be drawn over it. Stencil testing is only enabled for objects with
negative depth.

Next step is to only update the stencil buffer whenever the instance
changes since it's pointless to do it every frame

* scene: only update stencil buffer when needed

This is a bit jankier than I was hoping for and reminds me that all of
the stuff involving surface commits and synchronization really need to
be improved upon. It's a bit of a disaster, although not a huge issue in
practice (just annoying 1 frame artifacts when resizing the window)

* config, scene: add get_depth function

* config, scene: add depth field to object options

* scene: restore old default depth behavior

* scene: draw negative depth objects behind unsorted

* doc: document scene object depth

Added some documentation for scene objects themselves, along with the
new `depth` fields for images and mirrors.