server: do not send new wl_output modes
Sending new modes just causes a (small) memory leak in GLFW, and
Minecraft never makes use of the new modes anyways because its
fullscreen code is dumb as hell. Send one fake mode and leave it at that
so that we can properly deal with its attempts to "fullscreen" the game.