server: fix mouse in menus with ingame fullscreen
Due to an interaction between some perfectly sensible behavior in GLFW
and some extremely insensible behavior in Minecraft's fullscreen
handling code, this might be the only sane way to solve the bug where
enabling ingame fullscreen causes mouse movement in menus to break until
the window is resized.

GLFW will not send a window size change event if the window size didn't
actually change (shocker), but Minecraft immediately assumes that
whatever size it tried to set the window to is correct even though it
might not be. So, we need to actually change the size of the window to
get GLFW to tell Minecraft the window size has changed and it is wrong.

Closes #14