server: mark _NET_WM_STATE_FULLSCREEN as supported
LWJGL2's fullscreen code requires that _NET_WM_STATE_FULLSCREEN is
marked as supported by the WM, but doesn't ever use the state which is
supposed to be set on the window. So, marking it as supported is enough
for fullscreen to "work" with some rather large caveats, which I'd like
to fix later.