main, wrap, server: allow LWJGL2 fullscreen
The developers of LWJGL2 decided that (un)fullscreening the window
should involve destroying it and recreating it. I'm not really sure why,
but it's kind of a pain to deal with. I think this is the last straw for
convincing me to rework how fullscreen and windows in general are
handled, since there are now a lot of small annoying issues that have
added up due to the way things currently work.