main: setup environment passthrough
This is kinda jank but I'm becoming more accepting of jank after dealing
with fullscreen and LWJGL2 so whatever. This mostly fixes MangoHud (you
can launch the game with it, waywall still softlocks with it which
should be easily fixed though) and should make the hybrid GPU Nvidia
workaround much easier.