server: refactor pointer constraints
This implementation is also a bit more correct since it uses the
`already_constrained` error (at least for locked pointers) and works
based off of surfaces instead of whether the constraint is owned by the
client with input focus.

The main reason I did this was because I thought I broke something with
pointer constraints before (although the code is a bit better now,
I think, so it's not a waste of time) but it seems that at some point
river stopped activating pointer constraints until the surface is
committed. It seems plausible that other compositors do the same, so
I don't think it hurts to include the extra calls to wl_surface.commit.