util, treewide: improve string handling (#33)
* treewide: rename str -> strbuf

This makes the type name more in line with what it actually is (a
mutable buffer holding a string) and is in preparation for adding a new
str type which instead acts similar to Rust's str or C++'s
std::string_view in an effort to move away from null terminated strings.

* util: make strbuf a struct

Using an SDS-style string was interesting but I don't think the benefits
are worth the complexity. It also prevents _Generic from disambiguating
between strbuf and C strings.

* util: add str

This is intended to serve a similar purpose to Rust's str or C++'s
std::string_view.

* test: add tests for str, strbuf

* util: add str_clone_cstr, str_from

* env_reexec: use str for environment manipulation

* config: use str for action parsing

* util: add strbuf_clone_cstr

* server: use strbuf for X11 window title

* util: add ww_strdup

ww_strdup wraps strdup with a NULL check to avoid having to spam
check_alloc everywhere.

* treewide: use ww_strdup

* config, subproc, wrap: use strs for subprocesses

This removes the 63 argument limit (not that anyone was realistically
limited by it.)

* util: remove 0 capacity restriction on strbuf