lua: make ffi, load* available
The ffi package has been disabled for over a year and the load*
functions have been disabled for over a year and a half. There's not
really any point trying to fully sandbox the user configuration since it
requires some more invasive changes (require will still happily load
bytecode so it would be necessary to completely overwrite
package.loaders) and it will just get in the way of people trying to do
more interesting things with their configuration.